created site functions file
